Update Lecture.java
I changed getDate(void) to return a deep copy of the Date member variable to ensure immutability, and I documented all the methods. I also removed the Lecture(Date, FeedbackEntry) constructor because it is not useful to the extent that we would ever want to use it anywhere in our code. Such a constructor would make sense if we frequently had to make Lectures containing only a single FeedbackEntry, but that's not the case. Renamed member variable "ID" and method "getValues" to the more expressive "lectureNumber(void)" and "getRatingDistribution(void)",.