Add test ctor to FeedbackEntry.java
From what I understand, our encoded feedback list has the netID field mapped to an integer if it is valid and mapped to "null" if not valid. So in order to construct a FeedbackEntry without using an NRList to verify these entries, I added a constructor that only takes a String argument. There should be no other way to run tests on encoded netIDs without having a constructor like this (or using a dummy NRList to recognize the encoded netIDs). Could this be why LectureInitializer isn't working?