Generalized FeedbackEntry.main(), some formatting
I actually didn't mean to copy over main() from my system. But with some changes it can be useful for general testing. I allowed the user to select the feedback entry source file and optionally a roster source file (either as arguments to main() or through standard input). This is a good way to immediately test the integrity of any data sources (and to test the class itself).