Generalized PeerInteraction.main()
Made main() more general by allowing tester to pass in file name arguments for the feedback source and (optionally) the roster source. If arguments are not passed, the method will prompt the tester for that information. This permits convenient tests of the integrity of various data samples and also of the PeerInteraction class itself. I also recommend creating log files for at least the latest results of our tests so that we can see what isn't working.