Fixed RosterTest.java tests
File was out of date, leading to buggy ArrayIndexOutOfBoundsException(). However, the second test managed to catch an error with the Iterator on my machine, but only some of the time. So I updated the second test to run 100 times, just to be safe (probably not the best practice, but whatever). I'll fix Roster.java shortly.