Large overhaul of class descriptions
I restructured the documentation of all these classes in a manner I felt to be most logical. The main types of classes are (1) immutable, plain old data structures (NetIDPair and PeerInteraction) that provide essentially nothing but constructor and accessor methods, (2) associative containers that perform mapping operations (NRList and Roster), (3) the main objects of research, essentially Students and Lectures, which are very similar in terms of contents and statistics, (4) utilities used to generate graphs, restructure collections, fabricate test cases, etc. (Utilities, TextIO, and GraphTools), (5) the culmination of all these other elements into a finished (as far as Java code goes) product and a prototypical UI through which a developer can interact with all existing functionalities (ProtoApp and SimpleUI).