Methods fillQualCharacteristics and fillQuantCharacteristics polished off. However, testing cannot be ran on fillQuantCharacteristics due to absence of any quantitative data. Should not be an issue because the two methods entirely mimic each other apart from data type.
Modified Student class constructor so that it immediately initializes the quantitative and qualitative  HashMaps and then calls their respective 'fill' methods.

Also, these files will not work with main method and will fail all tests with main method because we currently lack files with encoded IDs in common. This is preventing us from full on testing integration. For this reason the StudentTest class was written in order to test the changes made to the Student class constructor and the methods to fill both HashMaps.